mysqldump.php now supports MySQL and SQLite (both tested), and (maybe) Postgres and MS-SQL (dblib PDO driver).
Extended to more databases: sqlite, Postegres and MS-SQL (dblib). Note that I only tested MySQL and sqlite! If you get a PDO Exception in one of the queries, it's probably because of a syntax error for the specific driver. To change that, go to the bottom of the file and edit the TypeAdapter class. For the good of mankind, don't hesitate to request a pull from your fork.