Change single-transaction so no more SUPER privilege is needed. Closes #54.
Previously, mysqldump-php modified global variables to guarantee transaction.
Now mysqldump-php only sets session variables, so no SUPER privilege is
needed.
start_transaction() has been splitted in two, (setup & start) so it should
be easier to implement a check to query the current transaction level. If
it is not REPEATABLE-READ, set it. Currenlty we are enforcing it.